requirements.txt
aiodns==3.0.0
aiohttp==3.8.4
aiohttp-json-rpc==0.13.3
aiosignal==1.3.1
apsw==3.41.2.0
async-timeout==4.0.2
attrs==23.1.0
Babel==2.12.1
black==23.3.0
Brotli==1.0.9
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contextlib2==21.6.0
discord.py==2.2.3
frozenlist==1.3.3
idna==3.4
importlib-metadata==6.6.0
Markdown==3.4.3
markdown-it-py==2.2.0
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
orjson==3.8.12
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.1
psutil==5.9.5
pycares==4.3.0
pycparser==2.21
Pygments==2.15.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
rapidfuzz==3.0.0
Red-Commons==1.0.0
Red-DiscordBot==3.5.2
Red-Lavalink==0.11.0
rich==13.3.5
schema==0.7.5
six==1.16.0
tomli==2.0.1
typing_extensions==4.5.0
yarl==1.9.2
zipp==3.15.0
